edited and feedback
some dragons edited completely and implemented the feedback system. if whoever is reading this, please send me an email if you need access to the feedback spreadsheet. I'd prefer to keep it private, but if you need it, I'd add you as a viewer to the sheet. 2738617@students.wits.ac.za